Confusion Matrix Metrics Plot

This function gives the user the ability to plot any of metrics computed from the confusion matrix over a grid of cut-off values for prediction going from 0.1 to 0.9 with steps of 0.1.

plot_metrics_over_cutoff(
  y,
  predicted_probabilities,
  plot_prevalence = FALSE,
  plot_accuracy = TRUE,
  plot_sensitivity = TRUE,
  plot_specificity = TRUE,
  plot_fdr = FALSE
)

Arguments

y

A vector containing n binary responses.

predicted_probabilities

A vector containing predicted responses using the estimated beta coefficients.

plot_prevalence

A boolean determining whether the prevalence metric is plotted or not. It has a default value of FALSE.

plot_accuracy

A numeric determining whether the accuracy metric is plotted or not. It has a default value of TRUE.

plot_sensitivity

A numeric determining whether the sensitivity metric is plotted or not. It has a default value of TRUE.

plot_specificity

A numeric determining whether the specificity metric is plotted or not. It has a default value of TRUE.

plot_fdr

A boolean determining whether the false discovery rate metric is plotted or not. It has a default value of FALSE.

Value

A plot of the selected metrics computed from the confusion matrix

Author

Ethan Sax, Nozara Sundus, Mohammad Maydanchi

Examples

#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#> Error in logistic_probability(beta_true, X): could not find function "logistic_probability"
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#> Error in eval(expr, envir, enclos): object 'logit_probs' not found
#Logistic regression to estimate coefficients
estimated_beta <- logistic_regression(X, y)
#> Error in eval(expr, envir, enclos): object 'y' not found
#Calculate the predicted probabilities using the estimated coefficients
predicted_prob <- logistic_probability(estimated_beta, X)
#> Error in logistic_probability(estimated_beta, X): could not find function "logistic_probability"
#Plotting the metrics over various cutoff values with user's choice
plot_metrics_over_cutoff(y, predicted_prob, plot_prevalence = FALSE,
plot_accuracy = TRUE, plot_sensitivity = TRUE,
plot_specificity = TRUE, plot_fdr = FALSE)
#> Error in eval(expr, envir, enclos): object 'predicted_prob' not found